Move instance group related code into separate file
File length for `role.go` kept growing. Separating instance
groups was an easy fix.
